* commit 'a22d198b1b7f3d6a61d916d408b1d52f8dd22343':
  Fix a cast in org.apache.harmony.tests.java.util.ArraysTest.

+++ harmony-tests/src/test/java/org/apache/harmony/tests/java/util/ArraysTest.java
@@ -2217,7 +2217,7 @@ public class ArraysTest extends junit.framework.TestCase {
                 long[] b = new long[a.length];
 
                 for (int i = 0; i < a.length; i++) {
-                    b[i] = (int) a[i];
+                    b[i] = (long) a[i];
                 }
                 return b;
             }
